does this work?

+++ mptestagent.py
@@ -159,10 +159,10 @@ class Player():
         if len(observation['Right'].shape)==\
             len(self.observation_space['Right'].shape):
             for name, obs in observation.items():
-                processed_obs[name] = obs[np.newaxis,:,:,:].astype(np.float32)/255
+                processed_obs[name] = tf.cast(obs[np.newaxis,:,:,:],tf.float32)/255
         else :
             for name, obs in observation.items():
-                processed_obs[name] = obs.astype(np.float32)/255
+                processed_obs[name] = tf.cast(obs, tf.float32)/255
         return processed_obs
 
     def choose_action(self, q):
